+++ src/java/entities/Medication.java
@@ -31,7 +31,7 @@ public class Medication implements Serializable{
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
     @NotNull
     private String medicationName;
     private String description;
@@ -55,7 +55,7 @@ public class Medication implements Serializable{
      * @param description
      * @param typeOfMedication
      */
-    public Medication(Integer id, String medicationName, String description, EnumMedType typeOfMedication) {
+    public Medication(Long id, String medicationName, String description, EnumMedType typeOfMedication) {
         this.id = id;
         this.medicationName = medicationName;
         this.description = description;
@@ -63,11 +63,11 @@ public class Medication implements Serializable{
     }
 
     //Getters & Setters
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
